Move GetRealLinkAndPlayTask to utils as top-level class
