+++ app/src/main/java/net/programmierecke/radiodroid2/RadioDroidBrowserService.java
@@ -6,7 +6,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.ServiceConnection;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.IBinder;
 import android.os.RemoteException;
@@ -17,58 +16,15 @@ import android.support.v4.media.MediaBrowserCompat;
 import android.support.v4.media.MediaBrowserServiceCompat;
 
 import net.programmierecke.radiodroid2.data.DataRadioStation;
+import net.programmierecke.radiodroid2.utils.GetRealLinkAndPlayTask;
 
-import java.lang.ref.WeakReference;
 import java.util.List;
-
-import okhttp3.OkHttpClient;
 
 public class RadioDroidBrowserService extends MediaBrowserServiceCompat {
     private RadioDroidBrowser radioDroidBrowser;
     private ServiceConnection playerServiceConnection;
     private IPlayerService playerService;
     private GetRealLinkAndPlayTask playTask;
-
-    public static class GetRealLinkAndPlayTask extends AsyncTask<Void, Void, String> {
-        private WeakReference<Context> contextRef;
-        private DataRadioStation station;
-        private WeakReference<IPlayerService> playerServiceRef;
-
-        private OkHttpClient httpClient;
-
-        public GetRealLinkAndPlayTask(Context context, DataRadioStation station, IPlayerService playerService) {
-            this.contextRef = new WeakReference<>(context);
-            this.station = station;
-            this.playerServiceRef = new WeakReference<>(playerService);
-
-            RadioDroidApp radioDroidApp = (RadioDroidApp) context.getApplicationContext();
-            httpClient = radioDroidApp.getHttpClient();
-        }
-
-        @Override
-        protected String doInBackground(Void... params) {
-            Context context = contextRef.get();
-            if (context != null) {
-                return Utils.getRealStationLink(httpClient, context.getApplicationContext(), station.ID);
-            }
-
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(String result) {
-            IPlayerService playerService = playerServiceRef.get();
-            if (result != null && playerService != null && !isCancelled()) {
-                try {
-                    playerService.SaveInfo(result, station.Name, station.ID, station.IconUrl);
-                    playerService.Play(false);
-                } catch (RemoteException e) {
-                    e.printStackTrace();
-                }
-            }
-            super.onPostExecute(result);
-        }
-    }
 
 
     private final BroadcastReceiver playStationFromIdReceiver = new BroadcastReceiver() {
